# backend/openapi/openapi-ext.yaml
openapi: 3.0.3
info:
  title: ОГЗ — расчёт погонажа МК (внешний API)
  description: >
    Внешний контракт калькулятора автоматического расчёта погонажа
    металлоконструкций и площади листовой стали для подготовки расчётов
    огнезащиты (ОГЗ). Версия спеки строго 3.0 (корп-требование).
  version: 0.1.0
servers:
  - url: /api/v1/ext
security:
  - bearerAuth: []
paths:
  /jobs:
    post:
      summary: Создать задание (загрузка PDF спецификации/чертежа)
      operationId: createJob
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                  description: PDF спецификации металла / чертёж металлокаркаса
      responses:
        "201":
          description: Задание создано
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobRef"
        "400":
          description: Файл отсутствует или это не PDF
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "413":
          description: Файл превышает допустимый размер
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    Unauthorized:
      description: Не авторизован
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
    JobRef:
      type: object
      properties:
        id:
          type: string
          format: uuid
        status:
          type: string
...